Revert "Cambios a la estructura de login"
