+++ package.json
@@ -48,7 +48,7 @@
     "prop-types": "^15.8.1",
     "storybook": "^8.3.0",
     "tailwindcss": "^3.4.11",
-    "vite": "^5.4.6"
+    "vite": "^5.4.1"
   },
   "eslintConfig": {
     "extends": [
