+++ src/App.jsx
@@ -4,15 +4,13 @@ import Login from './pages/LoginPage';
 import AttendancePage from './pages/AttendancePage';
 import PrivateRoute from './components/PrivateRoute';
 import 'mapbox-gl/dist/mapbox-gl.css';
-import LocationMap from './components/General/LocationMap';
-import GeolocationCheck from './components/General/GeolocationCheck'
 
 
 function App() {
     return (
         <Router>
             <Routes>
-                <Route path="/" element={<Login />} />
+                <Route path="/login" element={<Login />} />
 
                 {/* Ruta protegida para la página de asistencia */}
                 <Route
