+++ src/pages/LoginPage.jsx
@@ -2,92 +2,74 @@ import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { User, Lock } from 'lucide-react';
 import logo from '../assets/logo.jpg';
-import GeolocationCheck from '../Components/General/GeolocationCheck'; //Se importo el componente de GeolocationCheck -Andy
-import LocationMap from '../Components/General/LocationMap'; // Se importo el componente de LocationMap -Andy
 
 function Login() {
     const [username, setUsername] = useState('');
     const [password, setPassword] = useState('');
     const [loading, setLoading] = useState(false);
     const [error, setError] = useState('');
-    const [isLocationChecked, setIsLocationChecked] = useState(false);
-    const [userPosition, setUserPosition] = useState(null); // To store user's position
     const navigate = useNavigate();
 
     // Función para verificar si el usuario ya registró asistencia hoy
-const checkAttendance = async (userId) => {
-    const token = localStorage.getItem('token');
+    const checkAttendance = async (userId) => {
+        const token = localStorage.getItem('token');
 
-    try {
-        // Cambiamos las comillas por backticks para la interpolación de variables
-        const response = await fetch(`http://localhost:4000/api/attendance/today/${userId}`, {
-            headers: {
-                // También aquí corregimos la interpolación
-                Authorization: `Bearer ${token}`,
-            },
-        });
+        try {
+            const response = await fetch(`http://localhost:4000/api/attendance/today/${userId}`, {
+                headers: {
+                    Authorization: `Bearer ${token}`,
+                },
+            });
 
-        if (!response.ok) {
-            throw new Error('Error al verificar la asistencia');
+            if (!response.ok) {
+                throw new Error('Error al verificar la asistencia');
+            }
+
+            const attendanceData = await response.json();
+            return attendanceData; // Retorna los datos de asistencia
+        } catch (error) {
+            console.error('Error en la verificación de asistencia:', error);
+            setError('No se pudo verificar la asistencia.');
+            return null; // Retorna null en caso de error
         }
-
-        const attendanceData = await response.json();
-        return attendanceData; // Retorna los datos de asistencia
-    } catch (error) {
-        console.error('Error en la verificación de asistencia:', error);
-        setError('No se pudo verificar la asistencia.');
-        return null; // Retorna null en caso de error
-    }
-};
-
-// Función para registrar la asistencia (check-in)
-const registerAttendance = async (userId) => {
-    const token = localStorage.getItem('token');
-
-    // Obtener la fecha y hora actual en formato ISO
-    const checkInTime = new Date().toISOString();
-
-    // Establecer una ubicacion estatica o Dinamica
-    const location = userPosition; // Se hicieron cambios para que guarde la geolocation del check -Andy
-
-    try {
-        // Cambiamos las comillas por backticks para la interpolación de variables
-        const responseAttendance = await fetch(`http://localhost:4000/api/attendance`, {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json',
-                // También aquí corregimos la interpolación
-                Authorization: `Bearer ${token}`,
-            },
-            body: JSON.stringify({
-                user_id: userId,
-                check_in: checkInTime,
-                location: location,
-            }),
-        });
-
-        if (!responseAttendance.ok) {
-            throw new Error('Error al registrar la asistencia');
-        }
-
-        console.log('Asistencia registrada:', await responseAttendance.json());
-    } catch (error) {
-        console.error('Error en el registro de asistencia:', error);
-        setError('No se pudo registrar la asistencia.');
-    }
-};
-
-
-    // Funcion que maneja el output positivo de GeolocationCheck -Andy
-    const handleLocationSuccess = (position) => {
-        setIsLocationChecked(true); // Designa el check como positivo -Andy
-        setUserPosition(position); // Guarda la posicion del usuario -Andy
     };
 
-    // Funcion que maneja el geolocation fallido -Andy
-    const handleLocationFailure = () => {
-        setError('No estás dentro del área designada para iniciar turno.');
-        setIsLocationChecked(false); // Reinicia el location Check State
+    // Función para registrar la asistencia (check-in)
+    const registerAttendance = async (userId) => {
+        const token = localStorage.getItem('token');
+
+        // Obtener la fecha y hora actual en formato ISO
+        const checkInTime = new Date().toISOString();
+
+        // Establecer una ubicación estática o dinámica
+        const location = {
+            lat: 9.067326498950038, // Reemplaza con valores dinámicos si es necesario
+            lng: -79.38772760260268,
+        };
+
+        try {
+            const responseAttendance = await fetch(`http://localhost:4000/api/attendance`, {
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/json',
+                    Authorization: `Bearer ${token}`,
+                },
+                body: JSON.stringify({
+                    user_id: userId,
+                    check_in: checkInTime,
+                    location: location,
+                }),
+            });
+
+            if (!responseAttendance.ok) {
+                throw new Error('Error al registrar la asistencia');
+            }
+
+            console.log('Asistencia registrada:', await responseAttendance.json());
+        } catch (error) {
+            console.error('Error en el registro de asistencia:', error);
+            setError('No se pudo registrar la asistencia.');
+        }
     };
 
     const handleSubmit = async (e) => {
@@ -108,12 +90,12 @@ const registerAttendance = async (userId) => {
             setLoading(false);
 
             if (response.ok && data.accessToken) {
-                // Guarda los tokens el LocalStorage
+                // Guardar tokens en localStorage
                 localStorage.setItem('token', data.accessToken);
                 localStorage.setItem('refreshToken', data.refreshToken);
                 localStorage.setItem('role', data.user.role);
 
-                // Verificar si ya se registro asistencia hoy
+                // Verificar si ya registró asistencia hoy
                 const attendanceData = await checkAttendance(data.user.id);
 
                 if (attendanceData && !attendanceData.attendanceExists) {
@@ -140,76 +122,46 @@ const registerAttendance = async (userId) => {
     };
 
     return (
-        <div className="min-h-screen flex items-center justify-center bg-gradient-to-r from-blue-500 to-blue-700 relative">
-            {/* Contenedor para el error en la parte superior */}
-            {error && (
-                <div className="absolute top-8 w-full max-w-sm bg-white text-red-600 font-bold text-center py-2 px-4 rounded-md shadow-lg">
-                    <p>{error}</p>
-                </div>
-            )}
-            
+        <div className="min-h-screen flex items-center justify-center bg-gradient-to-r from-blue-500 to-blue-700">
             <div className="bg-white p-8 rounded-lg shadow-lg w-full max-w-sm">
                 <div className="flex justify-center mb-6">
                     <img src={logo} alt="Logo" className="w-24 h-24" />
                 </div>
-    
-                {/* Renderiza GeolocationCheck antes de mostrar el login form */}
-                {!isLocationChecked ? (
-                    <GeolocationCheck 
-                        desiredArea={{
-                            minLat: 9.061274114226507,   // Ajusta estos valores segun el area necesaria -Andy
-                            maxLat: 9.068140872874764,
-                            minLon: -79.39165233216502,
-                            maxLon: -79.38659580540434,
-                        }} 
-                        onSuccess={handleLocationSuccess} 
-                        onFailure={handleLocationFailure} 
-                    />
-                ) : (
-                    <>
-                        {userPosition && (
-                             <div style={{ height: '300px', width: '100%' }}>
-                                <LocationMap position={userPosition} />
-                            </div>        
-                        )} {/* Muestra el mapa si el Check Falla */}
-    
-                        <form onSubmit={handleSubmit}>
-                            <div className="mb-4 flex items-center">
-                                <User className="text-gray-700 mr-2" />
-                                <input
-                                    type="text"
-                                    placeholder="Username"
-                                    value={username}
-                                    onChange={(e) => setUsername(e.target.value)}
-                                    className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
-                                    required
-                                />
-                            </div>
-                            <div className="mb-6 flex items-center">
-                                <Lock className="text-gray-700 mr-2" />
-                                <input
-                                    type="password"
-                                    placeholder="Password"
-                                    value={password}
-                                    onChange={(e) => setPassword(e.target.value)}
-                                    className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
-                                    required
-                                />
-                            </div>
-                            <button
-                                type="submit"
-                                className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline w-full"
-                                disabled={loading}
-                            >
-                                {loading ? 'Iniciando sesión...' : 'Iniciar sesión'}
-                            </button>
-                        </form>
-                    </>
-                )}
+                <form onSubmit={handleSubmit}>
+                    <div className="mb-4 flex items-center">
+                        <User className="text-gray-700 mr-2" />
+                        <input
+                            type="text"
+                            placeholder="Username"
+                            value={username}
+                            onChange={(e) => setUsername(e.target.value)}
+                            className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+                            required
+                        />
+                    </div>
+                    <div className="mb-6 flex items-center">
+                        <Lock className="text-gray-700 mr-2" />
+                        <input
+                            type="password"
+                            placeholder="Password"
+                            value={password}
+                            onChange={(e) => setPassword(e.target.value)}
+                            className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+                            required
+                        />
+                    </div>
+                    {error && <p className="text-red-500 text-xs italic">{error}</p>}
+                    <button
+                        type="submit"
+                        className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline w-full"
+                        disabled={loading}
+                    >
+                        {loading ? 'Iniciando sesión...' : 'Iniciar sesión'}
+                    </button>
+                </form>
             </div>
         </div>
     );
-    
 }
 
 export default Login;
